feat: 삭제할 때 confirm 으로 사용자가 한번 더 확인 할 수 있게함

+++ src/controllers/eventHandlers/crewManageEvent.js
@@ -54,13 +54,14 @@ export function clickAddCrewButton(e) {
 
 export function clickDeleteCrewButton(e) {
   e.preventDefault();
-  console.log("click");
   const isCourseFrontEnd = document.getElementById("frontend-course").checked;
   const targetCrew = e.target.id;
-  if (isCourseFrontEnd) {
-    deleteCrew(FRONT_END_CREW_LIST_KEY, targetCrew);
-  } else {
-    deleteCrew(BACK_END_CREW_LIST_KEY, targetCrew);
+  if (confirm("정말 삭제하시겠습니까?")) {
+    if (isCourseFrontEnd) {
+      deleteCrew(FRONT_END_CREW_LIST_KEY, targetCrew);
+    } else {
+      deleteCrew(BACK_END_CREW_LIST_KEY, targetCrew);
+    }
   }
 }
